Deprecate TransactionGuard.submitTransaction usage
All the changes here are related to the deprecation of the
TransactionGuard.submitTransaction() methods related to this cherry
picked commit from JetBrains: http://go/androidcommit/162014c as part of
the 2020.1 merge.

For detailed information on how to adapt to the deprecated API, and an
explanation of the IDEA threading model (write-safe contexts, locks,
etc) please read this document: http://go/idea-threading-model

In some cases, one of the consequences of using invokeLater() instead
of submitTransaction() is that in tests, you might have to call
PlatformTestUtil.dispatchAllInvocationEventsInIdeEventQueue() to ensure
that the Runnables that are queued for execution on the EDT are
actually executed (since the tests run in a headless environment).
You can also call UIUtil.dispatchAllInvocationEvents() instead of
PlatformTestUtil.dispatchAllInvocationEventsInIdeEventQueue().
When using submitTransaction() in the past, this was not necessary.

Bug: 156102335

Change-Id: I6c782c70b7a5a047feec75af2ce7cfb0a6a8164c