Deprecate TransactionGuard.submitTransaction usage
All the changes here are related to the deprecation of the
TransactionGuard.submitTransaction() methods related to this cherry
picked commit from JetBrains: http://go/androidcommit/162014c as part of
the 2020.1 merge.

For detailed information on how to adapt to the deprecated API, and an
explanation of the IDEA threading model (write-safe contexts, locks,
etc) please read this document: http://go/idea-threading-model

1. forkResourceValue() runs on the EDT, and and is already in a
write-safe context (since the code runs as a result of user action in a
non-modal state). Also it is safe to read IDE data models from the EDT.
It calls createValueResource(), which uses a write command action in
order to mutate the IDE data models.

Please note that the following tests were failing before this change
and are still failing with this change.

- OverrideResourceActionTest.java (2 failures)

Bug: 156102335

Change-Id: I0b439383667c8d6bc6b54d4c46afc84303088107