Add uiMode and background to @Preview
Enable the use of repeated annotations to display multiple previews.

Now we allow users to specify uiMode as part of the preview annotation.
We also allow repeated @Preview annotations that allow, for example, to
use multiple uiMode annotations.

This way, we can have:

@Preview(name = "Light")
@Preview(name = "Dark", uiMode = Configuration.UI_MODE_NIGHT_YES)

Applied to the same element will generate both the light and dark
versions.

Test: New tests for the feature
Fixes: 155615344
Change-Id: I1aaaf10ba6ac6acf92b182ec70ad5156927cbed5