ProcessManagers should record devices with no usable processes.
We used to only record devices if there was at least 1 usuable
process. This causes the "Select Process" button to be empty
which is contrary to other similar runtime tools.

For the Legacy case we need to wait until the properties are
read. Otherwise we will record empty strings for manufacturer,
model, etc.

Fixes: 156787539
Test: Unit tests added & updated.

Change-Id: Ie2a6b382746a07c6f8cb583bd0b32a6d24e0e061