Validate and fix run configurations after sync
as they might get broken when opening a project previously synced
in the explicit module groups mode.

This process needs to run on sync completion when new modules are
available. It does not need to run on subsequent syncs but it should
be harmless considering a relatively small number of run configurations
present in a project usually.

Bug: 155359024
Test: OpenProjectIntegrationTest
Change-Id: Ibb963c72aedf91ca9b3c576f4dfc0afe86681a8d