[AUA] rewrite existing upgrade as refactorings
The RefactoringProcessor functionality is now sufficiently advanced to
express the logic found in AndroidPluginVersionUpdater.  Reimplement
the pieces there in terms of the refactoring processors.

(It may be that eventually the whole of AndroidPluginVersionUpdater is
replaced with the AgpUpgradeRefactoringProcessor, but that is not
sufficiently robust yet: having the implementation of
AndroidPluginVersionUpdater be in terms of parts of the implementation
of AgpUpgradeRefactoringProcessor will assist in the eventual
replacement, if that's desired.)

Bug: 149565531
Test: existing
Change-Id: Ia11d20e57e486a6ac8bdc18c4dd336bcd783d669