[AUA] invoke Upgrade Assistant from lint
If Lint as run within Studio detects an outdated dependency on the
Android Gradle Plugin, rather than accept only the simple text
substitution that Lint suggests as a fix, offer the user the action to
invoke the Upgrade Assistant to upgrade to the current recommended
version so that any associated refactorings or configuration changes
can also be triggered at the same time.

If the Lint check returns two fixes as alternatives, one is an actual
upgrade, and the other is an automatically-applicable micro-level
update; if it returns a single quick fix, it might be a significant
upgrade, or might be such a micro-level update.  Have the Upgrade
Assistant quickfix supersede any non-automatically-applicable
replacements, but allow micro-level updates to remain in the suggested
actions.

Bug: 157287524
Test: manual, existing tests updated
Change-Id: I0e9308c8a93ea19eb034013b47aa9a379defd926