Fix potential unordered access issue in app ID resolution
One of the underlying data structure in the ApplicationIdResolver
infrastructure had an unguarded read/write access being performed
by different threads.

This CL fixes this issue by making the caller class enforce
synchronization via the thread-safe container that contains the
data structure that was unsafely being accessed. The Java 8
compute* interface for Maps is used extensively since we take
advantage of the fact that ConcurrentHashMap guarantees
synchronization for these operations, in addition to its stream
interface.

Fixes: 157928072
Test: TBD
Change-Id: Icc21629667f03d9eed007be62414efa05b9b5746