Restore focus on table after loading is completed
If the table had focus before starting to load
we should restore the focus.
To prevent stealing the focus from other components
we give the focus to the progress bar when loading
starts. After loading is finished, if the table had
focus and the progress bad has focus, we restore
focus on the table.

Test: NA
Bug: 153209893
Change-Id: Ie3a2aa395384552566115e45908c45bad82297e8