Add option to only show references towards nearest GC root
This is an attempt to make the "Show nearest GC root"
feature more discoverable. In the future, when there are
other modes of viewing the "Reference" tab
(e.g. "Show all GC roots", "Omit cycles", etc.), we can
also turn the checkbox into a menu.

This CL also undoes a previous addition of this feature
as a context menu entry.

Bug: 157140322
Test: add new
Change-Id: I74f81aac591ecc9201b2dca8f201aae33a2689d7