[TP] Set traceId in batch queries to enforce correctness
Right now we're not setting the traceId for batch queries, which causes
TPD to just run the query against whatever trace is loaded at the
moment. Right now the possibility of a race condition while trying to
load two traces is almost none, but we should enforce it to make sure we
avoid any errors/mistakes and show wrong data.

Fixes: 157743759
Test: Updated TraceProcessorServiceImplTest + manual e2e
Change-Id: I4749bddbb3a0453caa3990aec99710922c16e591