Stabilize CreateCppKotlinProjectTest
There is a flakiness in this test that the menu
"File -> Sync Project with Gradle Files" sometimes
is not enabled. Call waitAndInvokeMenuPath() instead of
invokeMenuPath() and also give it a 20s time out.

Fixes: 158035256
Test: Ran this test with bazel 100 times, no failure.
Without this change, it failed around 8/100.
Also ran BuildCppKotlinTest to test waitAndInvokeMenuPath()
with default 10s time out.

Change-Id: Ife4eb3c99960eac16eaf19aeb074893c03b46bbd