Ignore exceptions in Gradle post-invoke tasks execution
The Gradle plugin has a CountDownLatch to indicate when the build
invocation is completed. Howevever, when a build is cancelled via
the status bar cancel button, this latch is never countDown()'ed.

This is caused by the fact that the task that counts down the
latch is in a post-invocation task, but the task is the second
task in the queue. When the user clicks on the cancellation
button, it causes all service manager getters to early out via
throwing a ProcessCanceledException. And the first task in the
queue does exactly this, therefore forcing an exit on the loop
that is processing all the post-invocation tasks.

I.e.:
for (Task task : tasks) {
  task.run(); // first task throws an exception, but our lock
              // is in the second task
}

Fixes: 158784803
Test: N/A
Change-Id: Ia24026e09d9b4b6a91b036df00fbf915afd7ca19