Fix leaks from Transforms and EasingCurve panels.
The inspector components are short lived and should clear up its
listeners to longer lived models when they "go away".
Unfortunatly there is no requirement that these panels implement
Disposable. Instead (or for now) do this on addNotify and
removeNotify.

Also: check for invisibility on the correct property.
Also: update the sliders with the value is null.

Bug: 159156848
Test: N/A
Change-Id: I9f8d7d03553398dbf2c1c6aada4ddaf87c677f56