Initial Compose pickers infra
Adds the APIs to add pickers based on PSI value/parameters and a *very*
basic implementation of the @Preview picker to exercise the API and
tests for it.

The CL includes the APIs to use the properties panel existing UI and
editors and map them to editing PSI properties in Kotlin code as opposed
 to XML
elements.

The basic picker provider just provides text editors for all properties
at the moment.
Future CLs will customize the editors to the specific types of the
properties.

Bug: 158041021
Test: Added tests for the PSI properties model.

Change-Id: I2624e8d0b773cb2df3a034f9f0664fff6fe7be8e