Add support for interpolated strings
Interpolated strings are interpreted in a special way. There are
multiple sub constants that are not handled by the
ConstantExpressionEvaluator.
This CL makes the evaluation manually and calculates the strings within
an interpolated strings.

For something like "Hello $name!!", two separate constants would be
generated, one for "Hello " and one for "!!".

Bug: 139475541
Test: New interpolation test
Change-Id: I90f4a799cb506e43773dd5874917197686609bd0