Stabilize AddNewBuildTypeTest
There is 1/1000 failure in AddNewBuildTypeTest:
The test was timeout when adding a new Build Variant.
From the screenshot, I see the IDE was loading "Fetching
Build models" running in the background.
In the test, I added waitForBackgroundTasks() before adding
new Build Variant and also increase the test time out from
3 mins to 5 mins.

Fixes: 158033006
Test: Ran this test 1k times, no failure
Change-Id: I92b371454e96b5c57fbfe43f7b5e86abeca9a34a