[AUA] incorporate into existing flow
Rather than overriding the existing AGP update window, have the new
Upgrade Assistant functionality come after it, still conditional on
the Studio Flag (so with the flag off, nothing changes).

For consistency, slightly change some of the wording of the existing
dialog if the Studio Flag is on: the title can use "Upgrade
Assistant"; use the word "Upgrade" rather than "Update" throughout,
and there is no need to mention the Gradle version update here (it
will be mentioned if necessary on the following dialog) or the new App
Bundle feature.

Bug: 149565531
Bug: 159612385
Test: existing
Change-Id: Id58127972aba9fce25230a724546bc4c2a9300d3