Enable blueprint mode for Compose previews
Under COMPOSE_BLUEPRINT_MODE flag (disabled by default).

This change simply enables the blueprint mode for Compose previews and
adds a toolbar button for it, similar to the one used in the XML Layout
Editor.

Note that this slightly changes the wording in the SceneMode button for
layout editor, most notable:
- 'Design + Blueprint' -> 'Design and Blueprint'

and the description:
- 'Show both Design and Blueprint' -> 'Show Design and Blueprint Surface'

Bug: 160021437
Test: manual
Change-Id: I045fdd97084f3c7e0fd85eaf9bcc56e9b081b743