Move google-cloud-tools-as earlier in the classpath
There are two classes:

./android-studio-plugin/resources/icons/GoogleCloudToolsIcons.java
./core-plugin/resources/icons/GoogleCloudToolsIcons.java

In the same packages, with similar and different fields.

Both are in the classpath, so we must keep the order right.

Note this is also the case when running tools/idea, but not
the final build as each plugin would have its own classloader.

We should ask google-cloud folks to fix this properly

Test: N/A
Bug: 156497102
Change-Id: I19a3a7d7aaa63709731bd041bba08cd1a0a669cb