Use new APIs in SymbolFilesLocator for symbol dirs
Also clean up some issues detected by IntelliJ.

Note on behavior change: with the new API, empty symbol directories and
skipped. So if user has only built one ABI, only the symbol dirs for
that ABI are collected.

Also, the logic now filters by variant name. Previously the filtering is
based on whether the variant contains "debug" in the name, and this does
not make sense to me.

Bug: 160258624
Test: manual & existing
Change-Id: I6682b6ac66ed536062ce2020981d4b6f0b9ad425