Migrate AppInspectorClient messenger API to kotlin coroutine
This CL converts AppInspectorClient.CommandMessenger to use suspend functions and converts DatabaseInspectorMessenger to use the new API. However, this CL does not attempt to completely port over DB inspector to coroutines, and it uses a sendCommandAsync function as temporary transition function between future and coroutine.

In addition, affected tests are converted to use coroutine.

Test: converted tests to use coroutine.
Bug: 144771043

Change-Id: I034c551cd09a0558afcc89c2fec5b65445b97125