Only one animation inspector can be open at a given time
If the animation inspector is open, all the toolbar buttons to open the
inspector, across all files containing Compose Previews, should be disabled.
The only exception is the button corresponding to the preview currently
being inspected. This button should be enabled so the user can click on it
to close the inspector. Closing the file where the animation inspector is
open should re-enable the toolbar button for all the previews.

Bug: 159189400
Test: ComposePreviewTest#testOnlyOneAnimationInspectorCanBeOpen
Change-Id: I4e98bdc0ecc4d761d61a349da5763c255d2f98b8