[safeargs] NavXmlIndex now supports custom tags
In navigation XML files, there is a concept of destinations - tags with
an ID, name, and a list of actions and arguments. There are explicit
destination tags -- activity, fragment, navigation, and dialog -- but
users can specify custom tags to.

Previously, we only supported explicit tags, forking a bunch of the
logic between them because, as far as navigation was concerned, they
were all just destinations.

After this CL, we split navigation tags into two partitions - ones we
explicitly look for -- action, argument, navigation -- and the rest,
which we just assume will always be a destination. We do this in a safe
way by creating a concept called a "MaybeDestination", which will always
succeed at being created / parsed if requested, but will only product a
destination instance if the proper values were set.

This is accomplished using the XmlAnyElement, a JAXB concept which you
are allowed to tag on one field per class, and, if present, is given the
opportunity to parse any encountered XML tags not otherwise handled.

Fixes: 158865326
Test: Updated / Covered by existing
Change-Id: I72cb9c792f2a4a516778b4a6384310f622ff95fd