[safeArg] No args should be passed if no destinations defined
This is a follow-up cl after initial fix to support args overrids.

1) In ag/12204194, we only fixed 'no destinations defined' case when
deciding whether to generate new inner action builder classes or
not. And this is not enough, as no args should be passed for static
action method as well.

So resolved actions are generated first and being used in all places.

2) Some code refactoring after 1).

Bug: 161353845
Test: Updated existing tests
Change-Id: Ia90f142a4a87a06c5373c0f83727bf45dbe39c91