[TPD] Track query metrics for TPD
Uses the new APIs for tracking TPD metrics inside
TraceProcessorService.

A pretty big re-work of the code was required, for two main reasons:

a) TraceProcessorService is an application-level service, unlike
   IdeProfilerServices.
   This means we cannot have TPDService depending on IdeProfilerServices
   in order to get the FeatureTracker, so we need to pass FeatureTracker
   around as we need to use it. This required a change in all APIs of
   TraceProcessorService.

b) TraceProcessorServiceImpl was very exception-driven, which made it
   very complex to catch/rethrow and properly collect metrics
   (query/method time) for all possible flows/cases.
   In order to improve that, TraceProcessorDaemonClient now always
   returns a result wrapper instead of throwing in case of problems and
   it's up to TraceProcessorServiceImpl to manage and interpret the
   result, even it means a query that was not completed.

As part of doing (b), this CL also fixes some mix uses of our
terminology, so now we should be clear that:
FAILURE -> Means we couldn't complete the query, as in the query failed
to reach TPD.
ERROR -> Means TPD received the query and managed to produce a response.
That response contained some error identified by TPD.

Bug: 147099016
Test: TraceProcessorServiceImplTest, PerfettoParserTest
Change-Id: Ia3caa3cf4c0e654e4cd935e395d1cb4f45f6c5c8