Re-work exception handling for correct parser
Currently, the logic in CpuCaptureParser will swallow all exceptions
thrown by the parsers, hiding some problems (like low buffer sizes for
gRPC in TraceProcessor backend) and making it harder to debug.

It was done like that, because we had to support the use case where the
user is loading a trace that we don't know the format and for some
parsers we don't have a clean way to check if the trace file is for that
particular technology (like we have for Atrace and Perfetto).

This CL changes the behaviour to, if we have a way to check the input
file and we already verified the input file is supported by that parser,
then any exception is treated as a real problem and we propagate it.

Test: Two new cases in CpuCaptureParserTest
Bug: 160916518
Change-Id: I34766fb676ec3e3be75c67828b914d65f6e13f53