[BA] Visualization rendered next to a tree
Implement visualization attached to a tasks tree as planned by BA redesign.
The idea is to have stack chart right next to the tasks tree with tree rows
connected to the corresponding sections. On scroll only tree rows should
scroll but stack should stay on place.

Implement is as a component attached to the right from the tree, both
components together go to the scroll pane.
Chart component uses tree model to get data for presentation,
requests tree row position on paint and uses currently visible
rect to calculate size and position of chart elements.

Bug: 158198410
Test: added
Change-Id: Iaa2ac7a49066cabc09e23441ebad50fd3d435f78