Clean up confusion between androidHome and SDK dir.
TL;DR:
ANDROID_HOME     => ANDROID_SDK_ROOT
ANDROID_SDK_HOME => ANDROID_PREFS_ROOT

Android Home served as the name of the preferences
folder (generally ~/.android) and could be overriden
via the property ANDROID_SDK_HOME

The SDK location was overrideable via the property
ANDROID_HOME.

This led to confusion in the code, naming of variable
that was not obvious (often containing Sdk to match
the property name instead of the content).

This commit cleans this up, using ANDROID_SDK_ROOT where
needed, sometimes in addition to ANDROID_HOME in order
to keep compatibility for a bit.
This does the same for ANDROID_SDK_HOME being replaced
by the new ANDROID_PREFS_ROOT.

Bug: N/A
Test: existing tests have been updated.
Change-Id: I25949714c19eafece8eb079c26e7f1237e3b2b25