On Run/Debug do not kill the app when installing
Currently, there is a race condition between the package manager
finish installing the app and terminating the process (force-stop)
and Studio launching the app ("am start"). Since the two are not
synced with respect to each other, a long running install (large
apk) may be delayed enough that when the package manager issues
the force-stop at the end of install, the stop happens after Studio
has issued the start command. This results in the app launching and
immediately getting killed.

So instead of syncing on the package manager (which is complicated)
we instead rely on the force-stop that happens right before deploy,
and instead tell the package manager to not stop the app upon end
of install (since we already stopped it from the Studio end).

Note that --dont-kill is available for N+ only. A followup will
address versions earlier than N.

Fixes: 160269604
Test: TBD, since this involves the package manager on device
Change-Id: I892c71c4c1601ca4c6b398ed321d923bc6268c45