Call updateSeekableAnimation() before setting the end state
For animations with multiple states, we set the end state to something
different than the initial state. A timing issue was happening for some
animations where we'd try to set this state before PreviewAnimationClock's
updateAnimationStates was called. As a result, the animation wouldn't
animate until a state change happened. This CL fixes that by making sure
this method is called before changing the end state.

Bug: 164509235
Test: ComposePreviewAnimationManagerTest
Test: Manually tested with StateAnimationWithInterruptionsDemo
Change-Id: Idf9b0920b10b33e0059a4f2ca5ea2182f7f5baab