Delete project system's Library sealed class
`Library`/`ExternalLibrary` classes/interfaces will be deprecated in
favor of the platform's project structure.

The existing implementation is inefficient in the Gradle project system
because of the data format mismatch. In the changes that follow we will
extract the interface of `ExternalLibrary` and make it possible for
project systems to provide custom efficient implementations.

Bug: 163607587
Test: n/a
Change-Id: Idb3fd156737c69dfe34b2a4beb46db686b8bc314