Simplify GradleTestArtifactSearchScopes initialization
Since AndroidModuleModel cannot appear by itself and is configured by
sync only there is no reason to attempt subsequent initialization of
scopes of the first attempt returned `null`. This makes it possible
for a scope to hold a non-null reference to its model and we can remove
some dubious `if`s.

Bug: 164730932
Test: n/a
Change-Id: I8d9444720a2ec198464a1ffdd008041eeacba464