Use IntelliJ Editor tabs style in Animation Inspector
The animation inspector tabs were using CommonTabbedPane, which is used
by othe Android Studio components, e.g. Profilers. However, this doesn't
result in a great UI when there are many tabs, as they end up being
vertically stacked.

After a UI/UX review, it was decided that the animation inspector tabs
should be similar to IntelliJ's editor tabs, being in a single row and
displaying a button that can be used to select hidden tabs when there are
too many.

Bug: 162960817
Test: ComposePreviewAnimationManagerTest + Manually checked visual changes.
Change-Id: I374bfdb0560fc81aba29b56955aae5e9dd8ba5a3