Fix inconsistent auto behaviour for API 29+ devices
The logic for invalidating the processes from a device was flawed.
- the invalidation could die out
- the collection used wasn't thread safe

When attaching to a new process we now retry if the client is
already connected, since we may be in the process of closing the
current client.

This change does not fix b/167579854 where a name of "Unknown" also
cause a failed auto start behaviour.

Fixes: 167420350
Test: Added unit test for the attach retry when connected
Change-Id: Iee5a6c2918de50d6bd1907fb4cb0da329dcfc7c2