Only run performance tests on its own rule
We need to remove the performance tests from the main test rule
(intellij.android.profilersAndroid_tests), because during each
postsubmit run we are effectively running the benchmark tests twice: the
main one and intellij.android.profilersAndroid.performance.test.

For memory consumption the values reported differs based on which test
rule it is being run at (because other code being ran in the same
suite/rule) and depending on the order the rules are ran we are overwriting
the perfgate jsons. Also under the main test rule there is a bigger variance in
the reported numbers.

This CL should make the performance numbers (at least the memory ones)
more stable.

Bug: n/a
Test: n/a
Change-Id: I20612128bc6b815d239a6fba47393c011d67522f